macOS: Don't set CMAKE_OSX_DEPLOYMENT_TARGET in CMakeLists
The minimum deployment target depends on whatever Qt version we build
against. Setting it in CMakeLists is not forward compatible and can
produce a bundle whose application binary and Info.plist declare a
lower minimum macOS version than the frameworks it ships with.

Turns out Qt 6.8 actually requires macOS 12 "Monterey", so bump the CI
deployment target to reflect this.